requirements.txt
numpy
opencv-python
vispy
PyQt6
pyserial
Pillow
